backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=24.0
reportlab>=4.0
python-dotenv>=1.0.0
Pillow>=10.0.0
